Shifted validate_ test cases to pytest
